Shift over execution specs to jobs
