+++ lib/good_job.rb
@@ -58,7 +58,7 @@ module GoodJob
 
   # @!attribute [rw] active_record_parent_class
   #   @!scope class
-  #   The ActiveRecord parent class inherited by +GoodJob::Execution+ (default: +ActiveRecord::Base+).
+  #   The ActiveRecord parent class inherited by +GoodJob::Job+ (default: +ActiveRecord::Base+).
   #   Use this when using multiple databases or other custom ActiveRecord configuration.
   #   @return [ActiveRecord::Base]
   #   @example Change the base class:
@@ -208,7 +208,7 @@ module GoodJob
     include_discarded = GoodJob.configuration.cleanup_discarded_jobs?
 
     ActiveSupport::Notifications.instrument("cleanup_preserved_jobs.good_job", { older_than: older_than, timestamp: timestamp }) do |payload|
-      deleted_executions_count = 0
+      deleted_jobs_count = 0
       deleted_batches_count = 0
       deleted_discrete_executions_count = 0
 
@@ -218,13 +218,11 @@ module GoodJob
         active_job_ids = jobs_query.pluck(:active_job_id)
         break if active_job_ids.empty?
 
-        if GoodJob::Execution.discrete_support?
-          deleted_discrete_executions = GoodJob::DiscreteExecution.where(active_job_id: active_job_ids).delete_all
-          deleted_discrete_executions_count += deleted_discrete_executions
-        end
+        deleted_discrete_executions = GoodJob::DiscreteExecution.where(active_job_id: active_job_ids).delete_all
+        deleted_discrete_executions_count += deleted_discrete_executions
 
-        deleted_executions = GoodJob::Execution.where(active_job_id: active_job_ids).delete_all
-        deleted_executions_count += deleted_executions
+        deleted_jobs = GoodJob::Job.where(active_job_id: active_job_ids).delete_all
+        deleted_jobs_count += deleted_jobs
       end
 
       if GoodJob::BatchRecord.migrated?
@@ -240,9 +238,9 @@ module GoodJob
 
       payload[:destroyed_batches_count] = deleted_batches_count
       payload[:destroyed_discrete_executions_count] = deleted_discrete_executions_count
-      payload[:destroyed_executions_count] = deleted_executions_count
+      payload[:destroyed_jobs_count] = deleted_jobs_count
 
-      destroyed_records_count = deleted_batches_count + deleted_discrete_executions_count + deleted_executions_count
+      destroyed_records_count = deleted_batches_count + deleted_discrete_executions_count + deleted_jobs_count
       payload[:destroyed_records_count] = destroyed_records_count
 
       destroyed_records_count
