+++ lib/good_job/current_thread.rb
@@ -13,7 +13,7 @@ module GoodJob
       error_on_discard
       error_on_retry
       error_on_retry_stopped
-      execution
+      job
       execution_interrupted
       execution_retried
       retry_now
@@ -49,11 +49,17 @@ module GoodJob
     #   @return [Exception, nil]
     thread_mattr_accessor :error_on_retry_stopped
 
-    # @!attribute [rw] executions
+    # @!attribute [rw] jobs
     #   @!scope class
     #   Execution
-    #   @return [GoodJob::Execution, nil]
-    thread_mattr_accessor :execution
+    #   @return [GoodJob::Job, nil]
+    thread_mattr_accessor :job
+
+    # # @!attribute [rw] executions
+    # #   @!scope class
+    # #   Execution
+    # #   @return [GoodJob::Execution, nil]
+    # thread_mattr_accessor :execution
 
     # @!attribute [rw] execution_interrupted
     #   @!scope class
@@ -92,7 +98,7 @@ module GoodJob
 
     # @return [String] UUID of the currently executing GoodJob::Execution
     def self.active_job_id
-      execution&.active_job_id
+      job&.active_job_id
     end
 
     # @return [Integer] Current process ID
