+++ spec/lib/good_job/active_job_extensions/interrupt_errors_spec.rb
@@ -30,32 +30,6 @@ RSpec.describe GoodJob::ActiveJobExtensions::InterruptErrors do
       )
     end
 
-    context 'when discrete execution is NOT enabled' do
-      before do
-        allow(GoodJob::Execution).to receive(:discrete_support?).and_return(false)
-      end
-
-      it 'is rescuable' do
-        TestJob.retry_on GoodJob::InterruptError
-
-        expect { GoodJob.perform_inline }.not_to raise_error
-        expect(GoodJob::Execution.count).to eq(2)
-
-        job = GoodJob::Job.first
-        expect(job.executions.order(created_at: :asc).to_a).to contain_exactly(have_attributes(
-                                                                                 performed_at: be_present,
-                                                                                 finished_at: be_present,
-                                                                                 error: start_with('GoodJob::InterruptError: Interrupted after starting perform at'),
-                                                                                 error_event: GoodJob::Job::ERROR_EVENT_RETRIED
-                                                                               ), have_attributes(
-                                                                                    performed_at: nil,
-                                                                                    finished_at: nil,
-                                                                                    error: nil,
-                                                                                    error_event: nil
-                                                                                  ))
-      end
-    end
-
     context 'when discrete execution is enabled' do
       before do
         allow(GoodJob::Execution).to receive(:discrete_support?).and_return(true)
@@ -66,11 +40,9 @@ RSpec.describe GoodJob::ActiveJobExtensions::InterruptErrors do
 
         expect { GoodJob.perform_inline }.not_to raise_error
         expect(GoodJob::Job.count).to eq(1)
-        expect(GoodJob::Execution.count).to eq(1)
         expect(GoodJob::DiscreteExecution.count).to eq(2)
 
         job = GoodJob::Job.first
-        expect(job.executions.count).to eq(1)
         expect(job.discrete_executions.count).to eq(2)
         expect(job).to have_attributes(
           performed_at: be_blank,
